Break out point, pixelref, pafcolor, options and displayparams to their own headers and delete a few related deprecated elements that are now only used in mgraph440
