+++ depthmapX/glpolygons.h
@@ -17,7 +17,7 @@
 #pragma once
 
 #include "genlib/p2dpoly.h"
-#include "salalib/vertex.h"
+#include "salalib/pafcolor.h"
 #include "depthmapX/gltrianglesuniform.h"
 #include <qopengl.h>
 #include <QVector>
