+++ salalib/salalib.pro
@@ -34,12 +34,13 @@ SOURCES += \
     sparksieve2.cpp \
     tigerp.cpp \
     topomet.cpp \
-    vertex.cpp \
     entityparsing.cpp \
     linkutils.cpp \
     gridproperties.cpp \
     geometrygenerators.cpp \
-    importutils.cpp
+    importutils.cpp \
+    point.cpp \
+    pafcolor.cpp
 
 HEADERS += \
     attributes.h \
@@ -60,7 +61,6 @@ HEADERS += \
     sparksieve2.h \
     tigerp.h \
     topomet.h \
-    vertex.h \
     entityparsing.h \
     linkutils.h \
     gridproperties.h \
@@ -68,7 +68,12 @@ HEADERS += \
     mgraph_consts.h \
     geometrygenerators.h \
     importutils.h \
-    importtypedefs.h
+    importtypedefs.h \
+    point.h \
+    pixelref.h \
+    displayparams.h \
+    pafcolor.h \
+    options.h
 
 DISTFILES += \
     salascript-tests.txt
